requirements.txt
backports.functools-lru-cache==1.5
certifi==2018.10.15
chardet==3.0.4
cycler==0.10.0
defusedxml==0.5.0
diff-match-patch==20121119
Django==1.9.8
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.1.0
et-xmlfile==1.0.1
fabio==0.8.0
future==0.16.0
h5py==2.9.0
httplib2==0.9.2
idna==2.7
jdcal==1.4
kiwisolver==1.1.0
matplotlib==2.2.4
numpy==1.15.3
odfpy==1.3.6
openpyxl==2.5.9
pandas==0.23.4
Pillow==5.3.0
pyparsing==2.4.0
python-dateutil==2.7.4
pytz==2018.6
PyYAML==3.13
requests==2.20.0
silx==0.9.0
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.24
web.py==0.39
xlrd==1.1.0
xlwt==1.3.0
